+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.2.0
 
++ Moved ApiStorageFile into admin module root folder as its now more then just a model.
++ Add file find() method to iterate trough files.
 + Added new methods for handling image filters.
 
 ## 1.1.0 (19. February 2019)